work with responses

+++ app/src/main/java/ru/zar1official/smartbathclient/data/models/BathState.kt
@@ -9,6 +9,6 @@ data class BathState(
     val craneActie: Boolean,
     val drainStatus: Boolean,
     val waterColor: Int,
-    val fillingProcent: Int,
+    val fillingProcent: Float,
     val timeStamp: Long
 )
